Create Common module and move common code to it
