+++ common/build.gradle
@@ -1,0 +1,2 @@
+group 'com.hartwig.healthchecks.common'
+version '1.0.0-SNAPSHOT'
