+++ common/src/main/java/com/hartwig/healthchecks/common/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package com.hartwig.healthchecks.common.exception;
+
+public class NotFoundException extends Exception {
+
+	public NotFoundException(String message) {
+		super(message);
+	}
+}
