--- src/main/java/com/hartwig/healthchecks/util/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package com.hartwig.healthchecks.util.exception;
-
-public class NotFoundException extends Exception {
-
-  public NotFoundException(String message) {
-    super(message);
-  }
-}
